requirements.txt
anthropic==0.51.0
openai==0.28.0
google-generativeai==0.8.3

numpy==2.2.6
pandas==2.2.3
matplotlib==3.10.3
seaborn==0.13.2
scikit_learn==1.6.1
scipy==1.15.3

openpyxl==3.1.5
protobuf==6.31.0

